+++ gym_duckietown/maps/loop_pedestrians.yaml
@@ -48,18 +48,9 @@ objects:
   rotate: 90
   height: 0.08
   static: False
-  pedestrian: True
 
 - kind: duckie
   pos: [4.0, 6.0]
   rotate: 90
   height: 0.08
   static: False
-  pedestrian: True
-
-- kind: duckie
-  pos: [3.5, 6.0]
-  rotate: 90
-  height: 0.08
-  static: False
-  pedestrian: True
