--- gym_duckietown/objects/pedestrian_obj.py
@@ -1,102 +1,0 @@
-import numpy as np
-from ..collision import *
-
-from .world_obj import WorldObj
-
-class PedestrianObj(WorldObj):
-    def __init__(self, obj, domain_rand, draw_bbox, road_tile_sz, dt=0.05):
-        super().__init__(obj, domain_rand, draw_bbox)
-        self.road_tile_sz = road_tile_sz
-
-        # Dynamic duckie stuff
-
-        # Randomize velocity and wait time
-        if self.domain_rand:
-            self.pedestrian_wait_time = np.random.randint(1, 30) * 10
-            self.vel = np.abs(np.random.normal(0.1, 0.02))
-        else:
-            self.pedestrian_wait_time = 100
-            self.vel = 0.02
-
-        # Movement parameters
-        self.heading = heading_vec(self.angle)
-        self.start = np.copy(self.obj['pos'])
-        self.center = self.obj['pos']
-        self.pedestrian_active = False
-        self.step_count = 0
-
-        # Wiggle2Walk parameter
-        self.wiggle = np.random.choice([10, 11, 12], 1)
-        self.wiggle = np.pi / self.wiggle
-
-    def check_collision(self, agent_corners, agent_norm):
-        """
-        See if the agent collided with this object
-        """
-        return intersects_single_obj(
-            agent_corners,
-            self.obj_corners.T,
-            agent_norm,
-            self.obj_norm
-        )
-
-    def safe_driving(self, agent_pos, agent_safety_rad):
-        """
-        See if the agent is too close to this object
-        """
-        d = np.linalg.norm(agent_pos - self.center)
-        score = d - agent_safety_rad - self.obj['safety_radius']
-
-        return min(0, score)
-
-    def step(self):
-        """
-        Use a motion model to move the object in the world
-        """
-        self.step_count += 1
-        self.pedestrian_active = np.logical_or(
-            self.step_count % self.pedestrian_wait_time == 0,
-            self.pedestrian_active
-        )
-
-        # If not walking, no need to do anything
-        if not self.pedestrian_active: 
-            return
-
-        # Update centers and bounding box
-        vel_adjust = self.heading * self.vel
-        self.center += vel_adjust
-        self.obj_corners += vel_adjust[[0, -1]]
-
-        distance = np.linalg.norm(self.center - self.start)
-
-        self.update_rendering()
-        if distance > self.road_tile_sz:
-            self.finish_walk()
-            
-
-    def update_rendering(self):
-        """
-        After moving, update any OpenGL attrib. for rendering
-        """
-        self.obj['pos'] = self.center
-        self.obj['y_rot'] = (self.angle + self.wiggle) * (180 / np.pi) 
-        self.wiggle *= -1
-        self.obj_norm = generate_norm(self.obj_corners)
-
-    def finish_walk(self):
-        """
-        After duckie crosses, update relevant attributes 
-        (vel, rot, wait time until next walk)
-        """
-        self.start = np.copy(self.center)
-        self.angle += np.pi
-        self.pedestrian_active = False
-
-        if self.domain_rand:
-            # Assign a random velocity (in opp. direction) and a wait time
-            self.vel = -1 * np.sign(self.vel) * np.abs(np.random.normal(0.1, 0.02))
-            self.pedestrian_wait_time = np.random.randint(1, 30) * 10
-        else:
-            # Just give it the negative of its current velocuty
-            self.vel *= -1
